backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
langchain>=0.3,<1.0
langchain-community>=0.3,<1.0
langchain-openai>=0.2,<1.0
langchain-text-splitters>=0.3,<1.0
chromadb>=0.5,<1.0
sentence-transformers>=3.0,<4.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
